pyproject.toml
[project]
name = "vidtracker"
version = "0.1.4"
description = "Video Object Tracking Toolkit"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Keyne Oei", email = "keynekassapa13@gmail.com" }]
requires-python = ">=3.8"
dependencies = [
    "contourpy>=1.1.1",
    "cycler>=0.12.1",
    "fonttools>=4.57.0",
    "importlib_resources>=6.4.5",
    "kiwisolver>=1.4.7",
    "loguru>=0.7.3",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "opencv-python>=4.11.0.86",
    "packaging>=25.0",
    "pillow>=10.4.0",
    "pyparsing>=3.1.4",
    "python-box>=7.2.0",
    "python-dateutil>=2.9.0.post0",
    "six>=1.17.0",
    "zipp>=3.20.2",
    "scikit-learn"
]

[project.scripts]
vidtracker = "vidtracker.cli:main"

[tool.setuptools]
packages = ["vidtracker"]

[tool.setuptools.package-data]
"vidtracker" = ["*.json"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
